+++ CHANGELOG.rst
@@ -3,10 +3,28 @@ Changelog
 =========
 
 All notable changes to this project will be documented in this file.
+The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0>`_, and
+this project tries its best to adhere to
+`Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0>`_,
-and this project adheres to `Semantic Versioning
-<https://semver.org/spec/v2.0.0.html>`_.
+Unreleased
+==========
+
+Added
+-----
+
+Changed
+-------
+- Documentation theme from Furo to the PyData-Sphinx-Theme.
+
+Deprecated
+----------
+
+Removed
+-------
+
+Fixed
+-----
 
 2023-01-25 - version 0.5.1
 ==========================